Collapse unused mixins into classes. Rename `virtual-repeater.js` to `VirtualRepeater.js`. (...)
`Repeats` and `RepeatsAndScrolls` were never actually used as mixins. All demos
used the `VirtualRepeater` and `VirtualScroller` classes, which apply these
mixins to empty classes (`class {}`).
