+++ preact/preact-repeater.js
@@ -1,5 +1,5 @@
 import {Component, h, render} from '../node_modules/preact/dist/preact.esm.js';
-import {VirtualRepeater} from '../src/virtual-repeater.js';
+import {VirtualRepeater} from '../src/VirtualRepeater.js';
 
 export class Repeat extends Component {
   componentDidMount() {
