+++ src/virtual-scroller.js
@@ -1,4 +1,4 @@
-import {Repeats} from './virtual-repeater.js';
+import {VirtualRepeater} from './VirtualRepeater.js';
 
 export class RangeChangeEvent extends Event {
   constructor(type, init) {
@@ -12,10 +12,9 @@ export class RangeChangeEvent extends Event {
   get last() {
     return this._last;
   }
-}
+};
 
-export const RepeatsAndScrolls = Superclass => class extends Repeats
-(Superclass) {
+export class VirtualScroller extends VirtualRepeater {
   constructor(config) {
     super();
 
@@ -440,5 +439,3 @@ export const RepeatsAndScrolls = Superclass => class extends Repeats
     }
   }
 };
-
-export const VirtualScroller = RepeatsAndScrolls(class {});
